Update troubleshooting.md (#67)
* Update troubleshooting.md

Question for reviewers: Should I move the information from the "conclusion" to the start of the file? To me this seems to be information that would be useful to know from the beginning. It could be integrated into the list of the three environments in the intro.

* Update troubleshooting.md

Removed the conclusion section, adding the information there to the bullet list at the beginning of the page.