Service/Ingress changes (#183)
* helper script to build and deploy a custom operator

* change service/ingress logic

Only create a LoadBalancer if a TCP port is exposed.

The ingress needs a nodeport or a loadbalancer. So we're replacing the
internal ClusterIP service with a NodePort service.

Also, the default port changed from 1 to whatever the target port is.
E.g. if the target port in the cluster is 1337, the public port will be
1337 too.

* delete challenges before deleting the cluster

* Automated commit: update images.

* print message that were sleeping on cluster delete

Co-authored-by: Stephen Roettger <stephen.roettger@gmail.com>