Revert back to w/o timeout
It turns out that `{timeout: 1500}` sets a timeout for the entire request to finish but throws when it is exceeded. This error would need to get caught and then decided upon what to do which complicates things. 

Changed the order so that page.goto can be awaited but still closed after a timeout.